+++ src/test/java/org/elasticsearch/index/engine/ShadowEngineTests.java
@@ -882,7 +882,7 @@ public class ShadowEngineTests extends ElasticsearchLuceneTestCase {
             wrapper.setRandomIOExceptionRate(randomDouble());
             wrapper.setRandomIOExceptionRateOnOpen(randomDouble());
             try (Store store = createStore(wrapper)) {
-                int refCount = store.refCount();
+                long refCount = store.refCount();
                 assertTrue("refCount: "+ store.refCount(), store.refCount() > 0);
                 Translog translog = createTranslog();
                 ShadowEngine holder;
